Lmh090626 png

+++ FinalProject/src/main/java/com/example/demo/project/boards/service/BoardsVO.java
@@ -3,9 +3,9 @@ package com.example.demo.project.boards.service;
 import lombok.Data;
 @Data
 public class BoardsVO {
-	private Integer id;
-	private Integer prjId;
+	private Long id;
+	private Long prjId;
 	private String name;
 	private String description;
-	private Integer topicsCount;
+	private Long topicsCount;
 }
