+++ FinalProject/src/main/java/com/example/demo/project/notice/service/NoticeService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface NoticeService {
 	List<NoticeVO> selectAll(NoticeVO notice);
-    NoticeVO selectOne(int id);
-    int delete(int id);
-    int update(NoticeVO notice);
-    int insert(NoticeVO notice);
+    NoticeVO selectOne(Long id);
+    Long delete(Long id);
+    Long update(NoticeVO notice);
+    Long insert(NoticeVO notice);
 }
